Moved some functions to install.js
